Add tension to lift at bottom; standardize units
As the robot's hardware gets closer to completion, we can finally start
to code again. Also, we found the gearing ratio! So rundown of changes:
If we get to the bottom, we go up a bit to not crash into a blooming
well of sadness. Aside from that just made units consistent and set the
spoolCircumference const.